feat: temporary file `usePonderRq` deleted and `usePonder` updated with dinamyc queries and `AllOffers` working.
